make compatible with TradedTokenContract by implementing IPresale. make timestamps uint64 instead uint256

+++ contracts/FundFactory.sol
@@ -126,9 +126,9 @@ contract FundFactory is Ownable, ReentrancyGuard, CostManagerFactoryHelper, Rele
      */
     function produce(
         address _sellingToken,
-        uint256[] memory _timestamps,
+        uint64[] memory _timestamps,
         uint256[] memory _prices,
-        uint256 _endTime,
+        uint64 _endTime,
         uint256[] memory _thresholds,
         uint256[] memory _bonuses
     ) 
@@ -166,9 +166,9 @@ contract FundFactory is Ownable, ReentrancyGuard, CostManagerFactoryHelper, Rele
     function produceToken(
         address _payToken,
         address _sellingToken,
-        uint256[] memory _timestamps,
+        uint64[] memory _timestamps,
         uint256[] memory _prices,
-        uint256 _endTime,
+        uint64 _endTime,
         uint256[] memory _thresholds,
         uint256[] memory _bonuses
     ) 
@@ -210,9 +210,9 @@ contract FundFactory is Ownable, ReentrancyGuard, CostManagerFactoryHelper, Rele
         address _sellingToken,
         address _token0,
         address _token1,
-        uint256[] memory _timestamps,
+        uint64[] memory _timestamps,
         uint256[] memory _prices,
-        uint256 _endTime,
+        uint64 _endTime,
         uint256[] memory _thresholds,
         uint256[] memory _bonuses
     )
